# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchiveMaster API",
        "description": "Sistema de gestión documental empresarial completo",
        "contact": {
            "name": "ArchiveMaster Support",
            "email": "admin@archivemaster.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "ArchiveMaster API Server"
        }
    ],
    "paths": {
...
        "/api/search/documents": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Búsqueda avanzada de documentos",
                "description": "Realizar búsqueda full-text y filtrada de documentos",
                "operationId": "db661f503942d2b8d19e57072f9388e4",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda (full-text)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "contrato servicios"
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filtrar por categoría",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "description": "Filtrar por creador",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "assigned_to",
                        "in": "query",
                        "description": "Filtrar por asignado",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    },
                    {
                        "name": "is_confidential",
                        "in": "query",
                        "description": "Filtrar por confidencialidad",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar por prioridad",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high"
                            ],
                            "example": "high"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Filtrar por tags (IDs separados por coma)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "1,2,3"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página (máximo 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "documents": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "document_number": {
                                                                "type": "string",
                                                                "example": "DOC-ABC-20250101120000-A1B2"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Contrato de Servicios"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Contrato de servicios profesionales"
                                                            },
                                                            "status": {
                                                                "type": "object"
                                                            },
                                                            "category": {
                                                                "type": "object"
                                                            },
                                                            "relevance_score": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 0.95
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total_results": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "search_time_ms": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "filters_applied": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search/users": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Búsqueda de usuarios",
                "description": "Buscar usuarios por nombre, email o posición",
                "operationId": "a3a196ba7169f33c764a3c56c7ea023c",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "juan perez"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filtrar por departamento",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "Filtrar por sucursal",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Límite de resultados (máximo 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "juan.perez@empresa.com"
                                                    },
                                                    "position": {
                                                        "type": "string",
                                                        "example": "Analista"
                                                    },
                                                    "department": {
                                                        "type": "object"
                                                    },
                                                    "branch": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar Bearer Token para autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}